Cadastros e edições funcionando, faltando edição baseadano curso

+++ projeto/projeto/curso_excluir.php
@@ -4,13 +4,28 @@
 
 	$id = $_GET['id'];
 
-	$sql = "DELETE FROM curso where id = $id";
+	$semAlunos = true;
+	
+    $sql = "select * from usuario";
+    $consulta = mysqli_query($conexao, $sql);
+    while($linha = mysqli_fetch_array($consulta)){
+    	if($linha['curso'] == $id) {
+    		$semAlunos = false;
+    	}
+    } // fim do while
 
-	$query = mysqli_query($conexao, $sql);
+	if ($semAlunos) {
+		$sql = "DELETE FROM curso where id = $id";
 
-	if($query){
-		header("location: curso.php?mensagem=sucesso");
+		$query = mysqli_query($conexao, $sql);
+
+		if($query){
+			header("location: curso.php?mensagem=sucesso");
+		}else{
+			header("location: curso.php?mensagem=erro");
+		}
 	}else{
-		header("location: curso.php?mensagem=erro");
-	}
+			header("location: curso.php?mensagem=erro");
+		}	
+?>
